pyproject.toml
[tool.poetry]
name = "mirgovorit"
version = "0.1.0"
description = ""
authors = ["Rustam Mukhtarov <mrustam9929@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
drf-yasg = "^1.21.7"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
